refactor(api): reorganize WebSocket adapter and improve layout styling
- Move webSocketAdapter from chat/ to api/adapters/ directory for better organization
- Update import path in MyRuntimeProvider to reflect new location
- Add dark background color to layout for consistent theming

This reorganization improves the API structure by grouping adapters in a dedicated directory.
