+++ src/lib/api/adapters/webSocketAdapter.ts
